Conflicts:
	Projects/SceneViewer/Classes/SelectSceneScreen.cpp
	Sources/Internal/Base/GlobalEnum.cpp
	Sources/Internal/CMakeLists.txt
	Sources/Internal/Core/Core.cpp
	Sources/Internal/Core/Core.h
	Sources/Internal/Render/Image/LibPVRHelper.cpp
	Sources/Internal/Render/PixelFormatDescriptor.cpp
	Sources/Internal/Render/RenderManager.h
	Sources/Internal/Render/RenderManagerGL.cpp
	Sources/Internal/Render/RenderState.cpp
	Sources/Internal/UI/UIMovieView.cpp

Former-commit-id: a1c92628f0cfbb5028ecc36ae617f4c0395552c7

+++ Tools/ResourceEditor/Classes/Qt/Scene/System/LandscapeEditorSystem.h
@@ -66,7 +66,7 @@ protected:
     
     
     DAVA::Texture* cursorTexture;
-    DAVA::uint32 cursorSize;
+    DAVA::float32 cursorSize;
 	Vector2 cursorPosition;
     Vector2 prevCursorPos;
